requirements.txt
# Core RAG dependencies
langchain>=0.3.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-huggingface>=0.1.0
langchain-groq>=0.1.9

# Embeddings (runs locally, no API key)
sentence-transformers>=3.0.0

# Vector store
chromadb>=0.5.0

# LLM (free Groq API)
groq>=0.9.0

# PDF support
pypdf>=4.0.0

# Web app
streamlit>=1.37.0

# Utilities
python-dotenv>=1.0.0
